+++ src/main/java/com/server/crews/auth/config/AuthArgumentResolverConfig.java
@@ -1,10 +1,11 @@
 package com.server.crews.auth.config;
 
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import java.util.List;
 
 @Configuration
 @RequiredArgsConstructor
