+++ src/main/java/com/server/crews/auth/repository/RefreshTokenRepository.java
@@ -7,5 +7,6 @@ import java.util.Optional;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, String> {
     void deleteByOwnerId(Long id);
+
     Optional<RefreshToken> findByToken(String token);
 }
