--- src/test/java/com/server/crews/auth/application/LoginServiceTest.java
@@ -1,107 +1,0 @@
-package com.server.crews.auth.application;
-
-import com.server.crews.auth.domain.Administrator;
-import com.server.crews.auth.domain.Applicant;
-import com.server.crews.auth.dto.request.AdminLoginRequest;
-import com.server.crews.auth.dto.request.ApplicantLoginRequest;
-import com.server.crews.auth.dto.response.AccessTokenResponse;
-import com.server.crews.auth.repository.AdministratorRepository;
-import com.server.crews.auth.repository.ApplicantRepository;
-import com.server.crews.environ.service.ServiceTest;
-import com.server.crews.recruitment.domain.Recruitment;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-
-import java.util.Optional;
-
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
-
-class LoginServiceTest extends ServiceTest {
-    @Autowired
-    private LoginService loginService;
-
-    @Autowired
-    private AdministratorRepository administratorRepository;
-
-    @Autowired
-    private ApplicantRepository applicantRepository;
-
-    @Test
-    @DisplayName("[동아리 관리자] 가입되지 않은 관리자가 로그인 요청을 하면 계정과 모집 공고를 생성하고 액세스 토큰을 발급한다.")
-    void loginNotSignedUpAdmin() {
-        // given
-        String email = "new@gamil.com";
-        String password = "new password";
-        AdminLoginRequest request = new AdminLoginRequest(email, password);
-
-        // when
-        AccessTokenResponse accessTokenResponse = loginService.loginForAdmin(request);
-
-        // then
-        Optional<Administrator> createdAdmin = administratorRepository.findById(accessTokenResponse.userId());
-        assertSoftly(softAssertions -> {
-            softAssertions.assertThat(createdAdmin).isNotEmpty();
-            softAssertions.assertThat(accessTokenResponse.accessToken()).isNotNull();
-        });
-    }
-
-    @Test
-    @DisplayName("[동아리 관리자] 가입된 관리자가 로그인 요청을 하면 액세스 토큰을 발급한다.")
-    void loginAdmin() {
-        // given
-        Administrator administrator = LIKE_LION_ADMIN().administrator();
-        AdminLoginRequest request = new AdminLoginRequest(administrator.getClubName(), administrator.getPassword());
-
-        // when
-        AccessTokenResponse accessTokenResponse = loginService.loginForAdmin(request);
-
-        // then
-        Optional<Administrator> createdAdmin = administratorRepository.findById(accessTokenResponse.userId());
-        assertSoftly(softAssertions -> {
-            softAssertions.assertThat(createdAdmin).isNotEmpty();
-            softAssertions.assertThat(accessTokenResponse.accessToken()).isNotNull();
-        });
-    }
-
-    @Test
-    @DisplayName("[동아리 지원자] 가입되지 않은 지원자가 로그인 요청을 하면 계정을 생성하고 액세스 토큰을 발급한다.")
-    void loginNotSignedUpApplicant() {
-        // given
-        Administrator publisher = LIKE_LION_ADMIN().administrator();
-        Recruitment recruitment = LIKE_LION_RECRUITMENT(publisher).recruitment();
-        String email = "new@gamil.com";
-        String password = "new password";
-        ApplicantLoginRequest request = new ApplicantLoginRequest(recruitment.getCode(), email, password);
-
-        // when
-        AccessTokenResponse accessTokenResponse = loginService.loginForApplicant(request);
-
-        // then
-        Optional<Applicant> createdApplicant = applicantRepository.findById(accessTokenResponse.userId());
-        assertSoftly(softAssertions -> {
-            softAssertions.assertThat(createdApplicant).isNotEmpty();
-            softAssertions.assertThat(accessTokenResponse.accessToken()).isNotNull();
-        });
-    }
-
-    @Test
-    @DisplayName("[동아리 지원자] 가입된 지원자가 로그인 요청을 하면 액세스 토큰을 발급한다.")
-    void loginApplicant() {
-        // given
-        Administrator publisher = LIKE_LION_ADMIN().administrator();
-        Recruitment recruitment = LIKE_LION_RECRUITMENT(publisher).recruitment();
-        Applicant applicant = JONGMEE_APPLICANT(recruitment).applicant();
-        ApplicantLoginRequest request = new ApplicantLoginRequest(recruitment.getCode(), applicant.getEmail(), applicant.getPassword());
-
-        // when
-        AccessTokenResponse accessTokenResponse = loginService.loginForApplicant(request);
-
-        // then
-        Optional<Applicant> createdApplicant = applicantRepository.findById(accessTokenResponse.userId());
-        assertSoftly(softAssertions -> {
-            softAssertions.assertThat(createdApplicant).isNotEmpty();
-            softAssertions.assertThat(accessTokenResponse.accessToken()).isNotNull();
-        });
-    }
-}
